refactor: move maxwell database schema management out
SchemaStore was becoming overly long.  Move it out to a different class.
